+++ JavaScriptRuntime/CommonJS/Require.cs
@@ -238,12 +238,17 @@ namespace JavaScriptRuntime.CommonJS
             var moduleDelegate = (ModuleMainDelegate)Delegate.CreateDelegate(typeof(ModuleMainDelegate), moduleEntryPoint);
 
             _currentParentModule = module;
+            
+            // Set the current require delegate so import() can access it
+            var previousRequire = RuntimeServices.SetCurrentRequire(moduleRequire);
             try
             {
                 moduleDelegate(module.exports, moduleRequire, module, canonicalId, dirName);
             }
             finally
             {
+                // Restore the previous require delegate
+                RuntimeServices.SetCurrentRequire(previousRequire);
                 _currentParentModule = parentModule;
                 module.MarkLoaded();
             }